Move legacy _get_task_inputs back to execution.py
Used only by UI
