+++ covalent_dispatcher/_core/execution.py
@@ -53,3 +53,36 @@ Defines the core functionality of the dispatcher
 #     _run_task,
 #     _run_task_and_update,
 # )
+
+from covalent._results_manager import Result
+
+from . import dispatcher, runner
+
+
+def _get_task_inputs(node_id: int, node_name: str, result_object: Result) -> dict:
+    """
+    Return the required inputs for a task execution.
+    This makes sure that any node with child nodes isn't executed twice and fetches the
+    result of parent node to use as input for the child node.
+
+    Args:
+        node_id: Node id of this task in the transport graph.
+        node_name: Name of the node.
+        result_object: Result object to be used to update and store execution related
+                       info including the results.
+
+    Returns:
+        inputs: Input dictionary to be passed to the task containing args, kwargs,
+                and any parent node execution results if present.
+    """
+
+    abstract_inputs = dispatcher._get_abstract_task_inputs(node_id, node_name, result_object)
+    input_values = runner._get_task_input_values(result_object, abstract_inputs)
+
+    abstract_args = abstract_inputs["args"]
+    abstract_kwargs = abstract_inputs["kwargs"]
+    args = [input_values[node_id] for node_id in abstract_args]
+    kwargs = {k: input_values[v] for k, v in abstract_kwargs.items()}
+    task_input = {"args": args, "kwargs": kwargs}
+
+    return task_input
